# api/openapi.yml
openapi: 3.1.0
info:
  title: LibreHousehold API
  description: "This API allows interaction with the backend of LibreHousehold, an open-source household management system."
  version: 1.0.0
  contact:
    name: LibreHousehold Maintainer Nicklas
    url: https://librehousehold.wiegandt.eu
    email: nicklas+librehousehold@wiegandt.eu
servers:
  - url: https://api.librehousehold.wiegandt.eu/v1
    description: The production server for LibreHousehold API
  - url: http://localhost
    description: Local development mokapi server
  - url: http://localhost:8080/v1
    description: Local started backend server
...
paths:
...
  /household/{householdId}/members:
    get:
      summary: Get all members of a household
      description: Retrieves all members of the specified household.
      tags:
        - members
      operationId: getMembers
      parameters:
        - name: householdId
          in: path
          required: true
          description: The unique identifier of the household.
          schema:
            type: string
            format: uuid
            example: "f3ce1a6b-aba1-478d-9f01-ec484f23ce60"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
        "404":
          description: No household found with the specified ID.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: An error occurred - see the status code and problem for more information.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      summary: Add a new member to a household
      description: Adds a new member to the specified household.
      tags:
        - members
      operationId: createMember
      parameters:
        - name: householdId
          in: path
          required: true
          description: The unique identifier of the household.
          schema:
            type: string
            format: uuid
            example: "f3ce1a6b-aba1-478d-9f01-ec484f23ce60"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Member"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          description: The data for the new member is invalid.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: No household found with the specified ID.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "409":
          description: The member already exists.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: An error occurred - see the status code and problem for more information.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/Problem'
  /household/{householdId}/members/{memberId}:
    get:
      summary: Get a specific member of a household
      description: Retrieves a specific member of the specified household.
      tags:
        - members
      operationId: getMember
      parameters:
        - name: householdId
          in: path
          required: true
          description: The unique identifier of the household.
          schema:
            type: string
            format: uuid
            example: "f3ce1a6b-aba1-478d-9f01-ec484f23ce60"
        - name: memberId
          in: path
          required: true
          description: The unique identifier of the member.
          schema:
            type: string
            format: uuid
            example: "7c3a40a8-3f57-4645-8cf5-820a6a8d3613"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: No household or member found with the specified IDs.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: An error occurred - see the status code and problem for more information.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      summary: Update a member's profile
      description: Updates the display name and/or email address of a household member. An email change is forwarded to the auth module via a domain event.
      tags:
        - members
      operationId: updateMember
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdate'
      responses:
        "204":
          description: Member updated.
        "404":
          description: Member not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "409":
          description: The email address is already in use.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: An error occurred.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      summary: Remove a member from the household
      description: Removes the specified member from the household. Can only be performed by the household admin.
      tags:
        - members
      operationId: removeMember
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Member removed.
        "404":
          description: Member not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: An error occurred.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /invite/{token}:
    get:
      summary: Resolve invite token
      description: Returns household info for a valid, non-expired invite token.
      tags:
        - members
      operationId: resolveInvite
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      responses:
        "200":
          description: Token resolved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteInfo'
        "404":
          description: Token not found or expired.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: An error occurred.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /invite/{token}/join:
    post:
      summary: Join a household via invite token (local registration)
      description: >
        Public endpoint. Creates an AccountEntity and a MemberEntity atomically,
        then establishes a session - mirrors POST /v1/auth/register for the join case.
      tags:
        - members
      operationId: joinHousehold
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocalMemberRegistration'
      responses:
        "201":
          description: Member created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        "404":
          description: Token not found or expired.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "409":
          description: E-mail address is already registered.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: An error occurred.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /household/join:
    post:
      summary: Complete household join after social login
      description: >
        Authenticated-only endpoint. Attaches the already-authenticated account
        (provisioned via OAuth2 JIT provisioning) as a member of the household
        referenced by the invite token. Mirrors POST /household/setup for the
        social-continuation half of the setup flow.
      tags:
        - members
      operationId: joinHouseholdAuthenticated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdJoin'
      responses:
        "201":
          description: Member created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        "404":
          description: Token not found or expired.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "409":
          description: The authenticated account is already a member of a household.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: An error occurred.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
    Problem:
      #$ref: 'https://opensource.zalando.com/restful-api-guidelines/models/problem-1.0.1.yaml#/Problem'
      type: object
      properties:
        type:
          type: string
          format: uri-reference
          description: >
            A URI reference that uniquely identifies the problem type only in the
            context of the provided API. Opposed to the specification in RFC-9457,
            it is neither recommended to be dereferenceable and point to a
            human-readable documentation nor globally unique for the problem type.
          default: 'about:blank'
          example: '/some/uri-reference'
        title:
          type: string
          description: >
            A short summary of the problem type. Written in English and readable
            for engineers, usually not suited for non technical stakeholders and
            not localized.
          example: some title for the error situation
        status:
          type: integer
          format: int32
          description: >
            The HTTP status code generated by the origin server for this occurrence
            of the problem.
          minimum: 100
          maximum: 600
          exclusiveMaximum: true
        detail:
          type: string
          description: >
            A human readable explanation specific to this occurrence of the
            problem that is helpful to locate the problem and give advice on how
            to proceed. Written in English and readable for engineers, usually not
            suited for non technical stakeholders and not localized.
          example: some description for the error situation
        instance:
          type: string
          format: uri-reference
          description: >
            A URI reference that identifies the specific occurrence of the problem,
            e.g. by adding a fragment identifier or sub-path to the problem type.
            May be used to locate the root of this problem in the source code.
          example: '/some/uri-reference#specific-occurrence-context'
...
    Member:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the member.
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: The name of the member.
        email:
          type: string
          format: email
          description: The email address of the member's account.
          example: "john.doe@example.com"
        avatar:
          type: string
          format: base64
          description: A base64 encoded image representing the member's avatar.
        isAdmin:
          type: boolean
          description: Indicates if the member is an admin of the household.
      example:
        id: "7c3a40a8-3f57-4645-8cf5-820a6a8d3613"
        name: "John Doe"
        avatar: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAACgAAAAoCAYAAACM/rhtAAAFZ0lEQVR4AcyYSUgcQRSG/5lDAlkIISSBbIQEErJdAlnIcgiEJAQEL4IigogoHvSgB0VFRRA9CYKgKOrJi7gvIN4EFRU9iTsq4oqi4IKIosavoIdxMs60M2YRXnfNq/f+/+uuKg/tHBgYOJqZmTlaWFj4rwIm2Jw3btzQ2tqarl69qjt37vwXAQtMsDkvXLigJ0+eaGJiQltbW/rXfzDAAhNsToAgJsEEBeT+ReANAywwwWAAGZBgggIKyf3NwBNvGGCxvF2AJJiggEIayP2NwAtPvGFw9zwByAQFFNJAI7k/GXjghSfenl6/AVJAIQ00IkDuTwTaeOCFpzcPr4AU0kAjAgiRO89AE2088DpN+1RAGmhEACEEyZ1HoIUm2nj40vQJSCMCCCGIMLlgAg200ETbn5ZfQAQQQhBhDMgFEvSigRaadjRsASKEIMIYYETuLEEPvWigZbfXNiCCCGOAEYbk7AS19NCLhp0eq+ZMgDRhgBGGGJPzFdRQSw+9vmq9zZ0ZEBGMMMQYAHLegjlqqKXHW42/XECAiGKIMQCAkHMPcsxRQ6373FnGAQNigjEAgABEjmBMjjlqyAUaQQFiCgAgAAFGMCbHHDXBRNCAmAMCEGAEY3LMBRvnAhgshK/+cwF0X1beHm+RnC9ju3NBAwICEGAsK8GYHHN2QU6rCwoQAEAAAswyYUyOOWqsfCD3gAExBgAQgDzNyTFHDbWe83Z/BwSIIcYAACJ5t2OOGmrp8V7lO3tmQIwwxBgA3/IyHwSopYdef/We82cCxAAjDO3AWWbU0kMvGlbezv1UwJ2dHXV1deno6MjoIIzB/fv3dfwdRwcHByZv9wIkvTU1Ndrc3DzRhtfY2JhXzVMBU1JS9PnzZ+3t7ZlPIsA1Nzfr9u3bevbsma5fv67BwcETRv5+ZGdnKzY2VsPDw0aT+ry8PF2+fPlUTa+ATU1NKi0tpV/b29vmuw333NxcVVVVaXl5WSEhIfr586f29/dNnb+Lu6a13J2dncrMzPSp+Rvg4uKioqKiFBcXZzx5cwg2Njbq48ePio6ONm8xLS1NKysrZhtkZGTo27dvrrdSXl6uL1++aHV11Wh4al65csV8sCorK9P79++9aprG48sJwMPDQ1McHh6u0NDQ42kZIfYPy/Lp0yeT4/L48WNump6eVkJCgnp7e5Wenq7x8XHzcPTfvHlT7poRERGmhwuarMTTp09dD+auSQ1xArCoqEizs7NiKefm5pgXT8tgdHTU7DvGxKVLl3Tr1i2z4e/du6eKigoVFxfr69ev+v79uxITEymTpVlYWGh+u19YnUePHpktxCF017TqXIDHXzSVnJyssLAw1dXViaejqKOjQ0tLS3rx4oVZUnJWsMQsP7/pe/78uebn55Wfny+n0ylLMzIyUj09Perr66NU7pobGxtmlYAF0l2TYhcgh+D169dqaGhQZWWlqqurmVdOTo6GhobMXmH5TPL4AsjxTS9fvuSm2tpajYyMmLfKaeXfE5pv3rxRa2urOQzoUuypyXLzoPxbY97SZOwCfPjwodi07CX+fVhiPDkH4N27d2prazNPzzZISkoSy/PgwQNxCGJiYpSamipOa0tLizgor169Un9/vyt8aa6vr6ukpER37949sZUMIK+WV8xT8DSQOxwObq748eOHgGJ/8TDd3d0GxuFwCDgORFZWlnnTgMbHx2tqasrVz8Dh8K3JFqivr9fk5KTr4Dj5R+wJh9iHDx/EMl28eJGfcjgcZsPzcZulZI9aS9He3m5g2OQUFxQUmF7rVJIj7Gi+ffvWtSdhc2LI13TeIkvlL3Z3d3Xt2jVzcPzV2p331IQFJth+AQAA//9Opj3vAAAABklEQVQDAM9TAgNdiTbvAAAAAElFTkSuQmCC"
...
    InviteInfo:
      type: object
      required:
        - householdId
        - householdName
        - validUntil
      properties:
        householdId:
          type: string
          format: uuid
          example: "f3ce1a6b-aba1-478d-9f01-ec484f23ce60"
        householdName:
          type: string
          example: "Die Müllers"
        validUntil:
          type: string
          format: date
          example: "2026-06-07"
    LocalMemberRegistration:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Max Mustermann"
        email:
          type: string
          format: email
          description: Email address used to create the joining member's local account.
          example: "max.mustermann@example.com"
        password:
          type: string
          minLength: 8
          example: "supersecret123"
        avatar:
          type: string
          format: base64
          example: "data:image/jpeg;base64,/9j/4AAQ..."
    HouseholdJoin:
      type: object
      required:
        - token
        - memberName
      properties:
        token:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        memberName:
          type: string
          minLength: 1
          maxLength: 100
          example: "Max Mustermann"
        memberAvatar:
          type: string
          format: base64
          example: "data:image/jpeg;base64,/9j/4AAQ..."
    MemberUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Jane Doe"
        email:
          type: string
          format: email
          example: "jane.doe@example.com"
...